Make prototype setup more straight-forward
