+++ lib/importer/index.js
@@ -1,53 +1,35 @@
-let in_memory_db = new Map();
+const multer = require("multer");
+const session = require("./session.js");
 
-const crypto = require("crypto");
+exports.Initialise = (config, router) => {
+  const storage = multer.diskStorage({
+    destination: (req, file, cb) => {
+      var fs = require("fs");
+      if (!fs.existsSync(config.uploadPath)) {
+        fs.mkdirSync(config.uploadPath, { recursive: true });
+      }
 
-exports.CreateSession = (request) => {
-  const response = {
-    id: "",
-    error: undefined,
-  };
+      cb(null, config.uploadPath);
+    },
 
-  const file = request.file;
-  if (!file) {
-    response.error = "Please attach a file";
-    return response;
-  }
+    filename: (req, file, cb) => {
+      cb(null, file.filename + "-" + Date.now());
+    },
+  });
 
-  const err = validateUpload(file);
-  if (err != undefined) {
-    response.error = err;
-    return response;
-  }
+  const uploader = multer({ storage });
 
-  // We need to tie the session id to a specific upload in such a way that
-  // we can find the upload from the session id, and also find session ids
-  // that we might consider timed out. For now we will use the hash of the
-  // filename and remember it in an in-memory dictionary
-  response.id = getFilenameHash(file.filename);
-  in_memory_db.set(response.id, file.filename);
+  router.post("/upload", uploader.single("file"), (request, response) => {
+    s = session.CreateSession(request);
 
-  return response;
+    if (s.error) {
+      response.render("upload.html", { error: s.error });
+      return;
+    }
+
+    response.render("success.html");
+  });
 };
 
-var getFilenameHash = (filename) => {
-  var hash = crypto.createHash("md5");
-  data = hash.update(filename, "utf-8");
-  var gen_hash = data.digest("hex");
-  return gen_hash;
-};
-
-var validateUpload = (file) => {
-  if (
-    file.mimetype !=
-    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
-  ) {
-    return "Uploaded file was not an XLSX file";
-  }
-
-  return undefined;
-};
-
-exports.ListSessions = () => {
-  return in_memory_db.keys();
-};
+exports.CreateSession = session.CreateSession;
+exports.ListSessions = session.ListSessions;
