+++ prototypes/basic/app/routes.js
@@ -2,20 +2,9 @@
 // For guidance on how to create routes see:
 // https://prototype-kit.service.gov.uk/docs/create-routes
 //
-const upload = require("./upload.js");
 const importer = require("importer");
 const govukPrototypeKit = require("govuk-prototype-kit");
 const router = govukPrototypeKit.requests.setupRouter();
+const cfg = require("govuk-prototype-kit/lib/config");
 
-const uploader = upload.uploader;
-
-router.post("/upload", uploader.single("file"), (request, response) => {
-  session = importer.CreateSession(request);
-
-  if (session.error) {
-    response.render("upload.html", { error: session.error });
-    return;
-  }
-
-  response.render("success.html");
-});
+importer.Initialise(cfg.getConfig(), router);
