+++ prototypes/basic/package.json
@@ -8,7 +8,6 @@
     "@govuk-prototype-kit/common-templates": "2.0.1",
     "govuk-frontend": "5.4.1",
     "govuk-prototype-kit": "13.16.2",
-    "importer": "file:../../lib/importer",
-    "multer": "^1.4.5-lts.1"
+    "importer": "file:../../lib/importer"
   }
 }
